Backend/app/services/ai_engine/simple_sd15_engine.py
"""
Simplified SD15 Engine for GTX 1650 - No ControlNet Dependency

This version works without controlnet-aux package by using
cv2 (OpenCV) directly for Canny edge detection.
"""

import torch
import gc
import logging
import time
import random
import base64
import io
from typing import Dict, List, Optional, Any, Tuple
from pathlib import Path
from PIL import Image
import numpy as np

try:
    from diffusers import StableDiffusionImg2ImgPipeline, DDIMScheduler
    DIFFUSERS_AVAILABLE = True
except (ImportError, RuntimeError) as e:
    logger = logging.getLogger(__name__)
    logger.warning(f"Diffusers not available (DLL/Import error): {e}")
    DIFFUSERS_AVAILABLE = False

try:
    import cv2
    CV2_AVAILABLE = True
except ImportError:
    CV2_AVAILABLE = False

from .base_engine import BaseEngine, GenerationRequest, GenerationResult, EngineType
from .prompt_builder import PromptBuilder

logger = logging.getLogger(__name__)


class SimpleSD15Engine(BaseEngine):
    """
    Simplified SD15 Engine using standard img2img without ControlNet.
    
    This avoids the controlnet-aux dependency and uses basic
    OpenCV for any preprocessing if needed.
    """
    
    def __init__(self, config: Dict[str, Any]):
        super().__init__(config)
        
        if not DIFFUSERS_AVAILABLE:
            logger.warning("Diffusers not available, using mock generation")
            self.use_mock = True
        else:
            self.use_mock = False
        
        # Configuration for GTX 1650 4GB
        self.device = config.get('device', 'cuda' if torch.cuda.is_available() else 'cpu')
        self.resolution = 512  # Fixed for 4GB VRAM
        self.num_inference_steps = config.get('num_inference_steps', 25)
        self.guidance_scale = config.get('guidance_scale', 7.5)
        self.strength = config.get('strength', 0.6)
        
        # Model - SD 1.5 for lower VRAM usage
        self.model_id = "runwayml/stable-diffusion-v1-5"
        
        # Pipeline
        self.pipeline = None
        
        # Memory optimization flags
        self.enable_attention_slicing = config.get('enable_attention_slicing', True)
        self.enable_cpu_offload = config.get('enable_cpu_offload', True)
        
        # Initialize
        self.prompt_builder = PromptBuilder()
        
        # Load pipeline if possible
        if not self.use_mock:
            self._load_pipeline()
        
        logger.info(f"Initialized Simple SD15 Engine")
        logger.info(f"Device: {self.device}")
        logger.info(f"Mock mode: {self.use_mock}")
    
    def _get_engine_type(self) -> EngineType:
        return EngineType.SD15_CONTROLNET
    
    def _load_pipeline(self):
        """Load the Stable Diffusion pipeline with memory optimizations."""
        try:
            logger.info("Loading Stable Diffusion 1.5 pipeline...")
            
            # Use fp16 for GPU to save memory
            dtype = torch.float16 if self.device == 'cuda' else torch.float32
            
            self.pipeline = StableDiffusionImg2ImgPipeline.from_pretrained(
                self.model_id,
                torch_dtype=dtype,
                safety_checker=None,
                requires_safety_checker=False
            )
            
            # Move to device
            self.pipeline = self.pipeline.to(self.device)
            
            # Apply memory optimizations
            if self.enable_attention_slicing:
                self.pipeline.enable_attention_slicing()
                logger.info("✅ Enabled attention slicing")
            
            if self.enable_cpu_offload and self.device == 'cuda':
                self.pipeline.enable_sequential_cpu_offload()
                logger.info("✅ Enabled CPU offload")
            
            logger.info("✅ Pipeline loaded successfully")
            
        except Exception as e:
            logger.warning(f"Failed to load pipeline: {e}")
            logger.info("Switching to mock generation mode")
            self.use_mock = True
            self.pipeline = None
    
    async def health_check(self) -> bool:
        """Check if the engine is healthy."""
        if self.use_mock:
            return True
        return self.pipeline is not None
    
    def get_model_info(self) -> Dict[str, Any]:
        """Get model information."""
        return {
            "engine_type": "Simple SD15 Engine (GTX 1650 Optimized)",
            "model_id": self.model_id,
            "device": self.device,
            "resolution": f"{self.resolution}x{self.resolution}",
            "mock_mode": self.use_mock,
            "status": "Ready" if self.use_mock or self.pipeline else "Not ready",
            "features": [
                "Image-to-image generation",
                "GTX 1650 4GB VRAM optimized",
                "Memory efficient (fp16 + attention slicing)",
                "CPU offload support",
                "Interior design prompts"
            ]
        }
    
    def _generate_mock_image(self, prompt: str, input_image: Image.Image, seed: int) -> str:
        """Generate a mock image for testing."""
        # Create a simple variation of input image
        width, height = self.resolution, self.resolution
        
        # Resize input
        img = input_image.resize((width, height), Image.Resampling.LANCZOS)
        
        # Add some variation based on seed
        np_img = np.array(img)
        np_img = (np_img + seed % 50 - 25).clip(0, 255).astype(np.uint8)
        img = Image.fromarray(np_img)
        
        # Convert to base64
        buffer = io.BytesIO()
        img.save(buffer, format='JPEG', quality=85)
        image_base64 = base64.b64encode(buffer.getvalue()).decode('utf-8')
        
        return f"data:image/jpeg;base64,{image_base64}"
    
    def _generate_real_image(self, prompt: str, input_image: Image.Image, seed: int) -> Optional[str]:
        """Generate a real image using the pipeline."""
        try:
            if not self.pipeline:
                return None
            
            # Resize input image to 512x512
            input_image = input_image.resize((self.resolution, self.resolution), Image.Resampling.LANCZOS)
            
            # Set seed
            generator = torch.Generator(device=self.device).manual_seed(seed)
            
            # Generate
            with torch.no_grad():
                result = self.pipeline(
                    prompt=prompt,
                    image=input_image,
                    num_inference_steps=self.num_inference_steps,
                    guidance_scale=self.guidance_scale,
                    strength=self.strength,
                    generator=generator
                )
            
            generated_image = result.images[0]
            
            # Clear cache
            if torch.cuda.is_available():
                torch.cuda.empty_cache()
            
            # Convert to base64
            buffer = io.BytesIO()
            generated_image.save(buffer, format='JPEG', quality=90)
            image_base64 = base64.b64encode(buffer.getvalue()).decode('utf-8')
            
            return f"data:image/jpeg;base64,{image_base64}"
            
        except torch.cuda.OutOfMemoryError:
            logger.error("GPU OOM - try reducing resolution or batch size")
            if torch.cuda.is_available():
                torch.cuda.empty_cache()
            return None
        except Exception as e:
            logger.error(f"Generation failed: {e}")
            return None
    
    async def generate_img2img(self, request: GenerationRequest) -> GenerationResult:
        """Generate interior design transformations."""
        start_time = time.time()
        
        try:
            # Validate request
            if not request.primary_image:
                return GenerationResult(
                    success=False,
                    generated_images=[],
                    error_message="Primary image is required",
                    engine_used="simple_sd15"
                )
            
            # Build prompts
            try:
                pos_prompt, neg_prompt = self.prompt_builder.build_prompt(
                    furniture_style=request.furniture_style,
                    wall_color=request.wall_color,
                    flooring_material=request.flooring_material
                )
            except:
                # Fallback
                pos_prompt = f"Professional interior design, {request.furniture_style} style, {request.wall_color} walls, {request.flooring_material} flooring, high quality, detailed"
                neg_prompt = "blurry, low quality, distorted, ugly, bad anatomy"
            
            # Load input image
            input_image = Image.open(io.BytesIO(request.primary_image)).convert('RGB')
            
            # Generate 3 variations
            generated_images = []
            seeds = []
            
            for i in range(3):
                seed = random.randint(0, 2**32 - 1)
                seeds.append(seed)
                
                logger.info(f"Generating variation {i+1}/3 with seed {seed}")
                
                if self.use_mock:
                    image_url = self._generate_mock_image(pos_prompt, input_image, seed)
                    generated_images.append(image_url)
                    logger.info(f"Generated mock image {i+1}")
                else:
                    image_url = self._generate_real_image(pos_prompt, input_image, seed)
                    if image_url:
                        generated_images.append(image_url)
                        logger.info(f"Generated real image {i+1}")
                    else:
                        logger.warning(f"Failed to generate image {i+1}")
            
            # Check results
            if not generated_images:
                return GenerationResult(
                    success=False,
                    generated_images=[],
                    error_message="Failed to generate any images",
                    engine_used="simple_sd15"
                )
            
            # Calculate time
            inference_time = time.time() - start_time
            
            result = GenerationResult(
                success=True,
                generated_images=generated_images,
                engine_used="simple_sd15",
                model_version=self.model_id,
                inference_time_seconds=inference_time,
                seeds_used=seeds
            )
            
            mode_text = "mock" if self.use_mock else "real"
            logger.info(f"✅ Generated {len(generated_images)} {mode_text} images in {inference_time:.2f}s")
            
            return result
            
        except Exception as e:
            logger.error(f"Generation failed: {e}")
            return GenerationResult(
                success=False,
                generated_images=[],
                error_message=str(e),
                engine_used="simple_sd15"
            )
        finally:
            # Cleanup
            if torch.cuda.is_available():
                torch.cuda.empty_cache()
            gc.collect()
